+++ src/js/submitPurchaseForm.js
@@ -1,9 +1,9 @@
 import { parsePrice } from '../input/parseInput.js';
 import { purchaseLottos } from '../service/PurchaseService.js';
 import validatePrice from '../validation/validatePrice.js';
-import { resetError, showError } from './util/errorHandler.js';
-import disableButton from './util/disabledButton.js';
-import { $ } from '../util/selector.js';
+import { resetError, showError } from '../util/web/errorHandler.js';
+import disableButton from '../util/web/disabledButton.js';
+import { $ } from '../util/web/selector.js';
 import { showLottoCount } from './view/showLottoCount.js';
 import { showLottoTickets } from './view/showLottoTickets.js';
 import showWinningNumberForm from './view/showWinningNumberForm.js';
